Ingredient units model field was integer, which caused fractions to round

+++ menu/models.py
@@ -21,7 +21,7 @@ class Ingredient(models.Model):
     id = models.AutoField(primary_key=True)
     name = models.CharField(max_length=200)
     recipe = models.ForeignKey(Recipe)
-    amount = models.IntegerField()
+    amount = models.DecimalField(max_digits=5, decimal_places=2)
     unit = models.CharField(max_length=200)
     def __str__(self):
         return str(self.amount) + " " + self.unit + " " + self.name
